Add a badge format mixin for RHs
